requirements.txt
pytest
sphinx>=1.3b3
numpydoc
